+++ src/main/kotlin/org/shangahi/sellio_backend/api/controller/FavoriteProductController.kt
@@ -1,7 +1,7 @@
 package org.shangahi.sellio_backend.api.controller
 
 import io.swagger.v3.oas.annotations.tags.Tag
-import org.shangahi.sellio_backend.api.dto.FavoriteToggleRequest
+import org.shangahi.sellio_backend.api.dto.request.FavoriteToggleRequest
 import org.shangahi.sellio_backend.api.dto.response.FavoriteProductsResponse
 import org.shangahi.sellio_backend.api.swagger.FavoriteProductDocs
 import org.shangahi.sellio_backend.service.FavoriteProductService
