+++ src/main/kotlin/org/shangahi/sellio_backend/api/dto/request/FavoriteToggleRequest.kt
@@ -1,4 +1,4 @@
-package org.shangahi.sellio_backend.api.dto
+package org.shangahi.sellio_backend.api.dto.request
 
 import java.util.UUID
